Code/requirements.txt
fastapi==0.111.0
numpy
Pillow==10.4.0
tensorflow
tensorflow_intel
uvicorn==0.30.1
